Remove articleVersion to rethink the solution.
