--- src/AppBundle/Entity/ArticleVersion.php
@@ -1,36 +1,0 @@
-<?php
-/**
- * @package AppBundle\Entity
- */
-
-namespace AppBundle\Entity;
-
-/**
- * Class ArticleVersion
- */
-class ArticleVersion extends AbstractArticleModel
-{
-    /**
-     * @var Article
-     */
-    private $article;
-
-    /**
-     * @param Article $article Set article.
-     * @return $this
-     */
-    public function setArticle(Article $article)
-    {
-        $this->article = $article;
-
-        return $this;
-    }
-
-    /**
-     * @return Article
-     */
-    public function getArticle()
-    {
-        return $this->article;
-    }
-}
