+++ src/AppBundle/Resources/config/doctrine/Article.orm.yml
@@ -8,10 +8,59 @@ AppBundle\Entity\Article:
             id: true
             generator:
                 strategy: AUTO
-    oneToMany:
-        versions:
-            targetEntity: ArticleVersion
-            mappedBy: article
+
+    fields:
+        title:
+            type: string
+            length: 255
+        slug:
+            type: string
+            length: 255
+        content:
+            type: text
+        publishAt:
+            column: publish_at
+            type: datetime
+            nullable: true
+        published:
+            type: boolean
+        slideshow:
+            type: boolean
+    manyToOne:
+        category:
+            targetEntity: Category
+            inversedBy: articles
+            joinColumn:
+                name: category_id
+                referencedColumnName: id
+        createdBy:
+            targetEntity: User
+            joinColumn:
+                name: created_by_id
+                referencedColumnName: id
+        modifiedBy:
+            targetEntity: User
+            joinColumn:
+                name: modified_by_id
+                referencedColumnName: id
+        event:
+            targetEntity: Event
+            inversedBy: articles
+            joinColumn:
+                name: event_id
+                referencedColumnName: id
+        related:
+            targetEntity: Related
+            inversedBy: articles
+            joinColumn:
+                name: related_id
+                referencedColumnName: id
+        thumbnail:
+            targetEntity: Image
+            inversedBy: articles
+            joinColumn:
+                name: thumbnail_id
+                referencedColumnName: id
     manyToMany:
         tags:
             targetEntity: Tag
@@ -25,4 +74,9 @@ AppBundle\Entity\Article:
                 inverseJoinColumns:
                     tag_id:
                         referencedColumnName: id
-    lifecycleCallbacks: { }
+    lifecycleCallbacks:
+        prePersist:
+                - setModifiedAt
+                - setCreatedAt
+        preUpdate:
+                - setModifiedAt
