--- src/AppBundle/Resources/config/doctrine/ArticleVersion.orm.yml
@@ -1,18 +1,0 @@
-AppBundle\Entity\ArticleVersion:
-    type: entity
-    table: null
-    repositoryClass: AppBundle\Repository\ArticleVersionRepository
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    manyToOne:
-        article:
-            targetEntity: Article
-            inversedBy: versions
-            joinColumn:
-                name: article_id
-                referencedColumnName: id
-    lifecycleCallbacks: { }
